rename test folder and update test driver
